requirements.txt
paramiko
python-dotenv
pyjwt
scp

